+++ source/ParametersGenome.cpp
@@ -6,6 +6,7 @@ void ParametersGenome::initialize(Parameters *pPin)
 {
     pP=pPin;
     gInsertOverlay=false;
+    gInsertOverlayDeltaFile="-";
     
     if (gDir.back()!='/') {
         gDir += '/';
@@ -13,10 +14,10 @@ void ParametersGenome::initialize(Parameters *pPin)
     if (gInsertOutDir!="-" && gInsertOutDir.back()!='/') {
         gInsertOutDir += '/';
     };
-    if (gInsertOutMode!="Full" && gInsertOutMode!="Overlay") {
+    if (gInsertOutMode!="Full" && gInsertOutMode!="Overlay" && gInsertOutMode!="Delta") {
         ostringstream errOut;
         errOut << "EXITING because of FATAL PARAMETER ERROR: unrecognized option in --genomeInsertOutMode = " << gInsertOutMode << "\n";
-        errOut << "SOLUTION: use one of the allowed values for --genomeInsertOutMode : Full or Overlay\n";
+        errOut << "SOLUTION: use one of the allowed values for --genomeInsertOutMode : Full, Overlay or Delta\n";
         exitWithError(errOut.str(), std::cerr, pP->inOut->logMain, EXIT_CODE_PARAMETER, *pP);
     };